+++ examples/react/src/App.tsx
@@ -1,63 +1,10 @@
-import { useSettings } from '@uxshop/storefront-react/lib/hooks/use-settings'
-import { useSections } from '@uxshop/storefront-react/lib/hooks/use-sections'
-import { useUser } from '@uxshop/storefront-react/lib/hooks/use-user'
-import { Sections } from '@uxshop/storefront-react/lib/components/sections'
-import { useApps, useBlogCategories, useBlogPosts, UxVideo } from '@uxshop/storefront-react'
-import { UxTextImage } from '@uxshop/storefront-react'
-import { useShowcases } from '@uxshop/storefront-react/lib/hooks/use-showcases'
-import { useEffect } from 'react'
-import Html from './components/html'
-
-const components = {
-  html: Html
-}
-
-const get = async (id: string) => {
-  console.log(`result ${id}`, '')
-}
-
-function Header() {
-  useEffect(() => {
-    const get = async (id: string) => {
-      console.log('')
-    }
-    get('header')
-  }, [])
-  return <h1>Hello </h1>
-}
+// import { ShopService } from '@uxshop/storefront-core'
+import { useStore } from '@uxshop/storefront-react'
 
 function App() {
-  const setting = useSettings()
-  const sections = useSections()
-  const user = useUser({ email: 'diovani@dooca.com.br', password: 'Teste123' })
-
-  const dataVideo = {
-    title: 'teste',
-    description: 'teste',
-    url: 'https://youtu.be/YD5hBvZ9fFo'
-  }
-
-  const dataUxTextImage = {
-    title: 'testando título',
-    description: 'testando descrição',
-    buttonHref: 'https://www.dooca.com.br',
-    buttonLabel: 'criar loja virtual',
-    image: 'https://dooca.com.br/assets/img/home/venda-no-instagram-e-google.jpg',
-    side: 'left'
-  }
-
-  const showcase = useShowcases({ pagination: { page: 2, items: 3 } })
-
-  useEffect(() => console.table(showcase.edges))
-
-  return (
-    <div className="App">
-      <h1>Teste {user && user.data.email}</h1>
-      <UxVideo settings={dataVideo} />
-      <UxTextImage settings={dataUxTextImage} />
-      <Sections components={components} />
-    </div>
-  )
+  // const sections: any = ShopService
+  const shop = useStore()
+  return <div className="App"></div>
 }
 
 export default App
