--- src/components/shop-microdata/types.ts
@@ -1,5 +1,0 @@
-import { Shop } from '@uxshop/storefront-core/dist/modules/shop/ShopTypes'
-
-export interface ShopMicroDataProps {
-    data: Shop
-}
