+++ src/components/store-microdata/index.tsx
@@ -2,6 +2,6 @@ import * as React from 'react'
 import { SeoServiceFactory } from '@uxshop/storefront-core/dist/services/seo/SeoServiceFactory'
 import { ShopMicroDataProps } from './types'
 
-export function ShopMicroData({ data }: ShopMicroDataProps) {
-    return <script type="application/ld+json">{SeoServiceFactory.getInstance('shop', data)}</script>
+export function StoreMicroData({ data }: ShopMicroDataProps) {
+  return <script type="application/ld+json">{SeoServiceFactory.getInstance('shop', data)}</script>
 }
