+++ src/components/store-microdata/types.ts
@@ -1,0 +1,3 @@
+export interface ShopMicroDataProps {
+  data: any
+}
