+++ src/index.ts
@@ -12,7 +12,7 @@ export { useMenus } from './hooks/use-menus'
 export { useScripts } from './hooks/use-scripts'
 export { useCategory } from './hooks/use-category'
 export { useCategoryTree } from './hooks/use-category-tree'
-export { useShop } from './hooks/use-shop'
+export { useStore } from './hooks/use-store'
 export { useLandingPages } from './hooks/use-landing-pages'
 export { useBrands } from './hooks/use-brands'
 export { SectionLoader } from './components/section-loader'
@@ -25,7 +25,7 @@ export { BlogPostMicroData } from './components/blog-microdata/blog-post'
 export { ProductListMicroData } from './components/product-microdata/product-list'
 export { ProductItemMicroData } from './components/product-microdata/product-item'
 export { SomeProductsMicroData } from './components/product-microdata/some-products'
-export { ShopMicroData } from './components/shop-microdata'
+export { StoreMicroData } from './components/store-microdata'
 export { PageMicroData } from './components/page-microdata'
 export { UxVideo } from './components/ux-video'
 export { UxTextImage } from './components/ux-text-image'
